core: Fix build of core::sync::atomic for BPF target
Before this change, build of any crate for bpfel-unknown-none target, as
as well as build of rustc library with the following config:

```
[build]
target = ["x86_64-unknown-linux-gnu", "bpfel-unknown-none"]
```

was failing with the following errors:

https://gist.github.com/vadorovsky/c9a3892a85ea4425082d91a729ffcfb2

That's because all these functions and macros (`atomic_int`,
`atomic_add`, `atomic_sub`) were annotated with:

```
\#[cfg(target_has_atomic = "8")]
```

However, BPF target has atomic operations enabled only for 64-bit types.
But all these functions are used by thew whole spectrum of types (8-bit,
16-bit, 32-bit, 64-bit, 128-bit and pointers), so this change changes
the annotation to:

```
\#[cfg(any(
    target_has_atomic = "8",
    target_has_atomic = "16",
    target_has_atomic = "32",
    target_has_atomic = "64",
    target_has_atomic = "128",
    target_has_atomic = "ptr"
))]
```

THIS IS NOT A FULL FIX, there are more problems on bpf-linker and LLVM
side which need investigation and fixing.

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>